requirements.txt
asyncpg==0.18.3
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
databases==0.2.5
fastapi==0.33.0
gunicorn==19.9.0
h11==0.8.1
h2==3.1.1
hpack==3.0.0
hyperframe==5.2.0
idna==2.8
importlib-metadata==0.19
more-itertools==7.2.0
numpy==1.17.0
packaging==19.1
pluggy==0.12.0
py==1.8.0
pydantic==0.30
pyparsing==2.4.2
pytest==5.0.1
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.6
starlette==0.12.0
toml==0.10.0
typing-extensions==3.7.4
urllib3==1.25.3
uvicorn==0.8.6
wcwidth==0.1.7
websockets==7.0
wsproto==0.15.0
zipp==0.5.2
